Refactor: move Ray scheduler code into dedicated sglang.srt.ray package
Extract Ray-specific scheduler launching and actor code from engine.py
into a new sglang.srt.ray package for cleaner separation of concerns.
- Move SchedulerActor from managers/scheduler_actor.py to ray/scheduler_actor.py
- Move Ray launching logic to ray/launch.py with RaySchedulerInitResult subclass
- Simplify engine.py SchedulerInitResult into base class + MpSchedulerInitResult
